+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NVIDIA Container Toolkit Changelog
 
 ## v1.14.0-rc.3
+* Added support for generating OCI hook JSON file to `nvidia-ctk runtime configure` command.
+* Remove installation of OCI hook JSON from RPM package.
 
 ## v1.14.0-rc.2